+++ R/sen2cor.R
@@ -1,7 +1,7 @@
 #' @title Correct L1C products using Sen2Cor
 #' @description The function uses Sen2Cor to manually correct L1C products.
 #'  Standalone version of
-#'  [sen2cor](http://step.esa.int/main/third-party-plugins-2/sen2cor)
+#'  [sen2cor](http://step.esa.int/main/third-party-plugins-2/sen2cor/)
 #'  (version 2.8.0 or 2.5.5) is used.
 #' @param l1c_prodlist List of L1C product names to be corrected. They can be both
 #'  product names with full/relative path or only names of SAFE products (in this case, also
@@ -83,6 +83,15 @@
 #'  from Python (leaving a standing process running).
 #' @param overwrite Logical value: should existing output L2A products be overwritten?
 #'  (default: FALSE)
+#' @param kill_errored Logical: experimental feature allowing killing dead Sen2Cor
+#'  processes, so leaving `sen2cor()` continuing processing on the remaining 
+#'  products. Set to TRUE to activate it (default is FALSE).
+#'  This experimental feature is available only on Unix systems,
+#'  and requires package "tools" to be installed.
+#'  This option is not compatible with `timeout` (in case both are set, 
+#'  `kill_errored` will be ignored).
+#' @param overwrite Logical value: should existing output L2A products be overwritten?
+#'  (default: FALSE)
 #' @param .log_message (optional) Internal parameter
 #'  (it is used when the function is called by `sen2r()`).
 #' @param .log_output (optional) Internal parameter
@@ -134,6 +143,7 @@ sen2cor <- function(
   tiles = NULL, 
   parallel = FALSE, 
   timeout = 0,
+  kill_errored = FALSE,
   overwrite = FALSE,
   .log_message = NA, 
   .log_output = NA
@@ -208,6 +218,30 @@ sen2cor <- function(
     return(invisible(NULL))
   }
   
+  # Check kill_errored
+  if (all(kill_errored == TRUE, Sys.info()["sysname"] == "Windows")) {
+    print_message(
+      type = "warning",
+      "'kill_errored' is available only for Unix systems; skipping it."
+    )
+    kill_errored <- FALSE
+  }
+  if (all(kill_errored == TRUE, !requireNamespace("tools", quietly = TRUE))) {
+    print_message(
+      type = "warning",
+      "'kill_errored' can be used only if package \"tools\" is installed; ",
+      "skipping it."
+    )
+    kill_errored <- FALSE
+  }
+  if (all(kill_errored == TRUE, timeout > 0)) {
+    print_message(
+      type = "warning",
+      "'timeout' and 'kill_errored' conflict; 'kill_errored' will be ignored."
+    )
+    kill_errored <- FALSE
+  }
+  
   ## Cycle on each file
   # if parallel==TRUE, use doParallel
   n_cores <- if (is.numeric(parallel)) {
@@ -217,6 +251,7 @@ sen2cor <- function(
   } else {
     min(parallel::detectCores()-1, length(l1c_prodlist), 8) # use at most 8 cores
   }
+  # n_cores <- 1
   # if (parallel==FALSE | Sys.info()["sysname"] == "Windows" | n_cores<=1) {
   if (parallel==FALSE | n_cores<=1) {
     `%DO%` <- `%do%`
@@ -229,13 +264,15 @@ sen2cor <- function(
   }
   
   # cycle on eacjh product
+  
   l2a_prodlist <- foreach(
     i=seq_along(l1c_prodlist),
     .combine=c,
     .export = "mountpoint",
     .packages='sen2r'
   ) %DO% {
-    
+  # l2a_prodlist <- c()
+  # for (i in seq_along(l1c_prodlist)) {
     # redirect to log files
     if (n_cores > 1) { # nocov start
       if (!is.na(.log_output)) {
@@ -433,19 +470,88 @@ sen2cor <- function(
         sel_sen2cor_log_message <- if (!is.na(.log_message)) {
           tempfile(pattern = "sen2cor_log_message_", fileext = ".txt")
         } else {NA}
-        sys::exec_wait(
-          binpaths$sen2cor,
-          args = sen2cor_args,
-          std_out = sel_sen2cor_log_output,
-          std_err = sel_sen2cor_log_message,
-          timeout = timeout
-        )
+        if (kill_errored == TRUE) {
+          sel_sen2cor_pid_1 <- sys::exec_background(
+            binpaths$sen2cor,
+            args = sen2cor_args,
+            std_out = sel_sen2cor_log_output,
+            std_err = sel_sen2cor_log_message
+          )
+        } else {
+          sys::exec_wait(
+            binpaths$sen2cor,
+            args = sen2cor_args,
+            timeout = timeout,
+            std_out = sel_sen2cor_log_output,
+            std_err = sel_sen2cor_log_message
+          )
+        }
       } else {
-        system(
-          paste(binpaths$sen2cor, paste(sen2cor_args, collapse = " ")),
-          intern = Sys.info()["sysname"] == "Windows",
-          timeout = timeout
-        )
+        if (kill_errored == TRUE) {
+          system(
+            paste(binpaths$sen2cor, paste(sen2cor_args, collapse = " ")),
+            intern = FALSE,
+            wait = FALSE
+          )
+        } else {
+          system(
+            paste(binpaths$sen2cor, paste(sen2cor_args, collapse = " ")),
+            intern = Sys.info()["sysname"] == "Windows",
+            timeout = timeout
+          )
+        }
+      }
+      
+      if (kill_errored == TRUE) {
+        Sys.sleep(10) # wait until all processes started
+        # Retrieve the correct PID
+        # (sel_sen2cor_pid_1 is the PID of L2A_Process bash script,
+        # while we need the PID of the second launched python utility)
+        psaux_raw0 <- system(paste0("ps aux"), intern = TRUE)
+        psaux_raw1 <- psaux_raw0[grepl(sel_l1c, psaux_raw0)]
+        psaux_raw2 <- strsplit(psaux_raw1, " +")
+        sel_sen2cor_pids <- sapply(psaux_raw2, function(x){x[[2]]})
+        sel_sen2cor_pid_2 <- if (length(sel_sen2cor_pids) > 0) {
+          max(sel_sen2cor_pids)
+          # psaux_raw2[[which(selmax(sel_sen2cor_pids)_sen2cor_pids==max(sel_sen2cor_pids))]][2]
+        } else {
+          1E9 # dummy nonexistent PID 
+        }
+        # check every 10 seconds
+        while (tools::pskill(sel_sen2cor_pid_2, 0)) {
+          # check sel_sen2cor_pid_2 to be running (CPU > 0)
+          sel_sen2cor_cpu <- 0
+          n_check_cpu <- 0
+          while (sel_sen2cor_cpu == 0) {
+            if (n_check_cpu <= 5) { # maximum 5 consecutive checks
+              Sys.sleep(10)
+              n_check_cpu <- n_check_cpu + 1
+              psaux_raw3 <- system(paste0("ps aux"), intern = TRUE)
+              psaux_raw4 <- psaux_raw3[grepl(sel_l1c, psaux_raw3)]
+              psaux_raw5 <- strsplit(psaux_raw4, " +")
+              sel_sen2cor_cpu <- if (
+                length(psaux_raw5) > 0 &&
+                sum(sapply(psaux_raw5, function(x){x[2]==sel_sen2cor_pid_2})) > 0
+              ) {
+                as.integer(psaux_raw5[[
+                  which(sapply(psaux_raw5, function(x){x[2]==sel_sen2cor_pid_2}))
+                  ]][3])
+              } else {
+                0
+              }
+            } else {
+              print_message(
+                type = "message", date = TRUE,
+                "Sen2Cor on ", basename(sel_l1c)," was killed after ",
+                format(Sys.time() - sel_sen2cor_timestart, digits = 3)," ",
+                "because inactive."
+              )
+              tools::pskill(sel_sen2cor_pids)
+              sel_sen2cor_cpu <- 1E6 # to exit from WHILE cycle
+            }
+          }
+        }
+        tools::pskill(sel_sen2cor_pids)
       }
       
       print_message(
@@ -512,6 +618,7 @@ sen2cor <- function(
     } # nocov end
     
     sel_l2a
+    # l2a_prodlist <- c(l2a_prodlist, sel_l2a)
     
   } # end FOREACH cycle on each product
   if (n_cores > 1) { # nocov start